+++ src/components/Portfolio/PortfolioHolding/PortfolioHoldingAddDialog.tsx
@@ -3,14 +3,12 @@ import { StockSearchItem } from "@api/stock";
 import BaseDialog from "@components/BaseDialog";
 import SearchBar from "@components/SearchBar/SearchBar";
 import Button from "@components/common/Buttons/Button";
+import BigDatePicker from "@components/common/DatePicker/BigDatePicker";
 import { Icon } from "@components/common/Icon";
-import { IconButton, ThemeProvider, createTheme } from "@mui/material";
-import { DatePicker } from "@mui/x-date-pickers";
-import designSystem from "@styles/designSystem";
+import { IconButton } from "@mui/material";
 import { useState } from "react";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
-import { IconCalendar } from "../../common/IconCalendar";
 
 type Props = {
   isOpen: boolean;
@@ -66,7 +64,6 @@ export default function PortfolioHoldingAddDialog({ isOpen, onClose }: Props) {
       onClose={onDialogClose}>
       <Header>
         <Title>종목 추가</Title>
-        {/* TODO: fix size */}
         <IconButton onClick={onDialogClose}>
           <Icon icon="close" size={24} color="gray600" />
         </IconButton>
@@ -76,7 +73,6 @@ export default function PortfolioHoldingAddDialog({ isOpen, onClose }: Props) {
         <div>
           종목 검색 <span>*</span>
         </div>
-        {/* TODO: width */}
         <SearchBar
           variant="select"
           sx={{ width: "480px" }}
@@ -99,19 +95,10 @@ export default function PortfolioHoldingAddDialog({ isOpen, onClose }: Props) {
       <InputContainer>
         <InputBox>
           <label>매입 날짜</label>
-          <ThemeProvider theme={muiTheme}>
-            <DatePicker
-              value={newPurchaseDate}
-              onChange={(newVal) => setNewPurchaseDate(newVal)}
-              format="YYYY-MM-DD"
-              slotProps={{
-                textField: { placeholder: "매입 날짜" },
-              }}
-              slots={{
-                openPickerIcon: IconCalendar,
-              }}
-            />
-          </ThemeProvider>
+          <BigDatePicker
+            value={newPurchaseDate}
+            onChange={(newVal) => setNewPurchaseDate(newVal)}
+          />
         </InputBox>
 
         <InputBox>
@@ -276,100 +263,3 @@ const Input = styled.input`
     color: ${({ theme: { color } }) => color.neutral.gray400};
   }
 `;
-
-const muiTheme = createTheme({
-  components: {
-    MuiFormControl: {
-      styleOverrides: {
-        root: {
-          width: "352px",
-          height: "32px",
-          border: "none",
-        },
-      },
-    },
-    MuiSvgIcon: {
-      styleOverrides: {
-        root: {
-          width: "16px",
-          height: "24px",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-        },
-      },
-    },
-    MuiInputAdornment: {
-      styleOverrides: {
-        root: {
-          width: "24px",
-          height: "100%",
-          margin: "0",
-          display: "flex",
-          justifyContent: "center",
-          alignItems: "center",
-        },
-      },
-    },
-    MuiIconButton: {
-      styleOverrides: {
-        root: {
-          padding: "0",
-          margin: "0",
-        },
-      },
-    },
-    MuiButtonBase: {
-      styleOverrides: {
-        root: {
-          width: "100%",
-          height: "100%",
-        },
-      },
-    },
-    MuiInput: {
-      styleOverrides: {
-        root: {
-          width: "87px",
-        },
-      },
-    },
-    MuiInputBase: {
-      styleOverrides: {
-        root: {
-          "display": "flex",
-          "padding": "4px 8px",
-          "justifyContent": "center",
-          "width": "352px",
-          "height": "32px",
-          "font": designSystem.font.body3,
-          "color": designSystem.color.neutral.gray400,
-          "borderColor": designSystem.color.neutral.gray100,
-          "backgroundColor": designSystem.color.neutral.white,
-          "&:focus": {
-            borderColor: designSystem.color.primary.blue500,
-          },
-          "&:hover": {
-            borderColor: designSystem.color.neutral.gray100,
-          },
-        },
-        input: {},
-      },
-    },
-    MuiOutlinedInput: {
-      styleOverrides: {
-        root: {},
-        input: {
-          "width": "304px",
-          "height": "21px",
-          "padding": "0 0 0 0",
-          "font": designSystem.font.body3,
-          "color": designSystem.color.neutral.gray900,
-          "::placeholder": {
-            color: designSystem.color.neutral.gray700,
-          },
-        },
-      },
-    },
-  },
-});
